Bruk felles objectMapper i stedet for å opprette ny
