+++ src/main/java/no/nav/familie/inntektsmelding/forespørsel/tjenester/task/OpprettForespørselTask.java
@@ -16,32 +16,31 @@ import no.nav.familie.inntektsmelding.forespørsel.tjenester.ForespørselBehandl
 import no.nav.familie.inntektsmelding.koder.ForespørselStatus;
 import no.nav.familie.inntektsmelding.koder.Ytelsetype;
 import no.nav.familie.inntektsmelding.metrikker.MetrikkerTjeneste;
+import no.nav.familie.inntektsmelding.typer.dto.OppdaterForespørselDto;
 import no.nav.familie.inntektsmelding.typer.dto.OrganisasjonsnummerDto;
 import no.nav.familie.inntektsmelding.typer.dto.PeriodeDto;
 import no.nav.familie.inntektsmelding.typer.dto.SaksnummerDto;
 import no.nav.familie.inntektsmelding.typer.entitet.AktørIdEntitet;
-import no.nav.familie.inntektsmelding.utils.K9ObjectMapper;
 import no.nav.vedtak.felles.prosesstask.api.ProsessTask;
 import no.nav.vedtak.felles.prosesstask.api.ProsessTaskData;
 import no.nav.vedtak.felles.prosesstask.api.ProsessTaskHandler;
+import no.nav.vedtak.mapper.json.DefaultJsonMapper;
 
 @ApplicationScoped
 @ProsessTask("forespørsel.opprett")
 public class OpprettForespørselTask implements ProsessTaskHandler {
     private static final Logger LOG = LoggerFactory.getLogger(OpprettForespørselTask.class);
+    private static final ObjectMapper OBJECT_MAPPER = DefaultJsonMapper.getObjectMapper();
 
     public static final String YTELSETYPE = "ytelsetype";
     public static final String ORGNR = "orgnr";
     public static final String STP = "skjaeringstidspunkt";
 
     private ForespørselBehandlingTjeneste forespørselBehandlingTjeneste;
-    private ObjectMapper objectMapper;
 
     @Inject
-    public OpprettForespørselTask(ForespørselBehandlingTjeneste forespørselBehandlingTjeneste,
-                                  K9ObjectMapper k9ObjectMapper) {
+    public OpprettForespørselTask(ForespørselBehandlingTjeneste forespørselBehandlingTjeneste) {
         this.forespørselBehandlingTjeneste = forespørselBehandlingTjeneste;
-        this.objectMapper = k9ObjectMapper.getObjectMapper();
     }
 
     OpprettForespørselTask() {
@@ -78,13 +77,34 @@ public class OpprettForespørselTask implements ProsessTaskHandler {
         }
 
         try {
-            etterspurtePerioder = objectMapper.readValue(
+            etterspurtePerioder = OBJECT_MAPPER.readValue(
                 prosessTaskData.getPayloadAsString(),
-                objectMapper.getTypeFactory().constructCollectionType(List.class, PeriodeDto.class)
+                OBJECT_MAPPER.getTypeFactory().constructCollectionType(List.class, PeriodeDto.class)
             );
             return etterspurtePerioder;
         } catch (Exception e) {
             throw new RuntimeException("Kunne ikke deserialisere etterspurtePerioder for ytelse: " + ytelsetype, e);
         }
+    }
+
+    public static ProsessTaskData lagOpprettForespørselTaskData(Ytelsetype ytelsetype,
+                                                          AktørIdEntitet aktørId,
+                                                          SaksnummerDto saksnummer,
+                                                          OppdaterForespørselDto forespørselDto) {
+        var taskdata = ProsessTaskData.forProsessTask(OpprettForespørselTask.class);
+        taskdata.setProperty(OpprettForespørselTask.YTELSETYPE, ytelsetype.name());
+        taskdata.setAktørId(aktørId.getAktørId());
+        taskdata.setSaksnummer(saksnummer.saksnr());
+        taskdata.setProperty(OpprettForespørselTask.ORGNR,  forespørselDto.orgnr().orgnr());
+        taskdata.setProperty(OpprettForespørselTask.STP, forespørselDto.skjæringstidspunkt().toString());
+        if (forespørselDto.etterspurtePerioder() != null) {
+            try {
+                taskdata.setPayload(OBJECT_MAPPER.writeValueAsString(forespørselDto.etterspurtePerioder()));
+            } catch (Exception e) {
+                LOG.error("Kunne ikke serialisere etterspurtePerioder til JSON", e);
+                throw new RuntimeException("Kunne ikke serialisere etterspurtePerioder", e);
+            }
+        }
+        return taskdata;
     }
 }
